app/requirements.txt
Flask
mysql-connector
mysql-connector-python